requirements.txt
fastapi
python-dotenv
redis[hiredis]
sqlalchemy
ulid
mysqlclient
pymysql